requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.2
jupyter>=1.0.0
tqdm>=4.61.0
pyserial>=3.5